optimize the Base class and enhance request robustness

+++ imagedl/modules/sources/base.py
@@ -83,16 +83,14 @@ class BaseImageClient():
     def _removeduplicates(self, image_infos: list = None):
         unique_image_infos, identifiers = [], set()
         for image_info in image_infos:
-            if image_info['identifier'] in identifiers:
-                continue
+            if image_info['identifier'] in identifiers: continue
             identifiers.add(image_info['identifier'])
             unique_image_infos.append(image_info)
         return unique_image_infos
     '''_appenduniquefilepathforimages'''
     def _appenduniquefilepathforimages(self, keyword: str, image_infos: list):
         time_stamp = datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
-        work_dir = sanitize_filepath(os.path.join(self.work_dir, self.source, f'{time_stamp} {keyword.replace(" ", "")}'))
-        touchdir(work_dir)
+        work_dir = sanitize_filepath(os.path.join(self.work_dir, self.source, f'{time_stamp} {keyword.replace(" ", "")}')); touchdir(work_dir)
         for idx, image_info in enumerate(image_infos):
             image_info['work_dir'] = work_dir
             image_info['file_path'] = os.path.join(work_dir, f'{str(idx+1).zfill(8)}')
@@ -104,21 +102,14 @@ class BaseImageClient():
         while len(search_urls) > 0:
             search_url = search_urls.pop(0)
             search_inputs, search_url, request_method = (search_url['inputs'], search_url['url'], search_url['method']) if isinstance(search_url, dict) else ({}, search_url, 'get')
-            resp: requests.Response = getattr(self, request_method)(search_url, **search_inputs, **request_overrides)
-            if resp is None or resp.status_code != 200:
-                bar()
-                continue
+            try: resp: requests.Response = getattr(self, request_method)(search_url, **search_inputs, **request_overrides); resp.raise_for_status()
+            except Exception: bar(); continue
+            if (resp is None) or (resp.status_code not in {200}): bar(); continue
             resp.encoding = 'utf-8'
-            try:
-                search_result = self._parsesearchresult(resp.text)
-            except Exception as err:
-                self.logger_handle.error(f'{self.source}._search >>> {search_url} (Error: {err})', disable_print=self.disable_print)
-                search_result = []
-            if isinstance(search_result, dict):
-                image_infos.append(search_result)
-            else:
-                assert isinstance(search_result, list)
-                image_infos.extend(search_result)
+            try: search_result = self._parsesearchresult(resp.text)
+            except Exception as err: self.logger_handle.error(f'{self.source}._search >>> {search_url} (Error: {err})', disable_print=self.disable_print); search_result = []
+            if isinstance(search_result, dict): image_infos.append(search_result)
+            else: assert isinstance(search_result, (list, tuple)); image_infos.extend(list(search_result))
             bar()
         return image_infos
     '''search'''
@@ -135,15 +126,13 @@ class BaseImageClient():
         with alive_bar(len(search_urls)) as bar:
             for _ in range(num_threadings):
                 task = threading.Thread(target=self._search, args=(search_urls, bar, image_infos, request_overrides))
-                task_pool.append(task)
-                task.start()
+                task_pool.append(task); task.start()
             for task in task_pool: task.join()
         # logging
         image_infos = self._removeduplicates(image_infos)
         self._appenduniquefilepathforimages(image_infos=image_infos, keyword=keyword)
         if len(image_infos) > 0:
-            work_dir = image_infos[0]['work_dir']
-            touchdir(work_dir)
+            work_dir = image_infos[0]['work_dir']; touchdir(work_dir)
             self._savetopkl(image_infos, os.path.join(work_dir, 'search_results.pkl'))
         else:
             work_dir = self.work_dir
@@ -157,29 +146,18 @@ class BaseImageClient():
     '''_download'''
     @usedownloadheaderscookies
     def _download(self, image_infos: list, bar: alive_bar, request_overrides: dict = None, downloaded_image_infos: list = None):
-        request_overrides = request_overrides or {}
-        checked_work_dirs = set()
-        for image_info in image_infos:
-            if image_info['work_dir'] not in checked_work_dirs:
-                touchdir(image_info['work_dir'])
-                checked_work_dirs.add(image_info['work_dir'])
+        request_overrides, checked_work_dirs = request_overrides or {}, set()
+        for info in image_infos: (wd := info["work_dir"]) in checked_work_dirs or (touchdir(wd), checked_work_dirs.add(wd))
         while len(image_infos) > 0:
-            image_info = image_infos.pop(0)
-            file_path, image_candidate_urls = image_info['file_path'], image_info['candidate_urls']
+            image_info = image_infos.pop(0); file_path, image_candidate_urls, resp = image_info['file_path'], image_info['candidate_urls'], None
             for image_candidate_url in image_candidate_urls:
-                resp = self.get(image_candidate_url, **request_overrides)
-                if resp is not None and resp.status_code == 200: break
-            if resp is None or resp.status_code != 200: bar(); continue
-            assert (not os.path.exists(file_path))
+                try: (resp := self.get(image_candidate_url, **request_overrides)).raise_for_status(); break
+                except Exception: continue
+            if (resp is None) or (resp.status_code not in {200}) or os.path.exists(file_path): bar(); continue
             with open(file_path, 'wb') as fp: fp.write(resp.content)
-            ext = imghdr.what(file_path)
-            if ext in ["rgb", "gif", "pbm", "pgm", "ppm", "tiff", "rast", "xbm", "jpeg", "jpg", "bmp", "png", "webp", "exr"]:
-                file_path_with_ext = f'{file_path}.{ext}'
-                assert (not os.path.exists(file_path_with_ext))
-                shutil.move(file_path, file_path_with_ext)
-                downloaded_image_info = copy.deepcopy(image_info)
-                downloaded_image_info['file_path'] = file_path_with_ext
-                downloaded_image_infos.append(downloaded_image_info)
+            if (ext := imghdr.what(file_path)) in {"rgb", "gif", "pbm", "pgm", "ppm", "tiff", "rast", "xbm", "jpeg", "jpg", "bmp", "png", "webp", "exr"}:
+                shutil.move(file_path, f'{file_path}.{ext}')
+                downloaded_image_info = copy.deepcopy(image_info); downloaded_image_info['file_path'] = f'{file_path}.{ext}'; downloaded_image_infos.append(downloaded_image_info)
             else:
                 os.remove(file_path)
             bar()
@@ -196,19 +174,24 @@ class BaseImageClient():
         with alive_bar(len(image_infos)) as bar:
             for _ in range(num_threadings):
                 task = threading.Thread(target=self._download, args=(image_infos, bar, request_overrides, downloaded_image_infos))
-                task_pool.append(task)
-                task.start()
+                task_pool.append(task); task.start()
             for task in task_pool: task.join()
         # logging
         if len(downloaded_image_infos) > 0:
-            work_dir = downloaded_image_infos[0]['work_dir']
-            touchdir(work_dir)
+            work_dir = downloaded_image_infos[0]['work_dir']; touchdir(work_dir)
             self._savetopkl(downloaded_image_infos, os.path.join(work_dir, 'download_results.pkl'))
         else:
             work_dir = self.work_dir
         self.logger_handle.info(f'Finished downloading images using {self.source}. Download results have been saved to {work_dir}, valid downloads: {len(downloaded_image_infos)}.', disable_print=self.disable_print)
         # return
         return downloaded_image_infos
+    '''_autosetproxies'''
+    def _autosetproxies(self):
+        if self.auto_set_proxies:
+            try: self.session.proxies = self.proxied_session_client.getrandomproxy()
+            except Exception as err: self.logger_handle.error(f'{self.source}._autosetproxies >>> freeproxy lib failed to auto fetch proxies (Error: {err})', disable_print=self.disable_print); self.session.proxies = {}
+        else:
+            self.session.proxies = {}
     '''get'''
     def get(self, url, **kwargs):
         if 'cookies' not in kwargs: kwargs['cookies'] = self.default_cookies
@@ -218,21 +201,10 @@ class BaseImageClient():
             if not self.maintain_session:
                 self._initsession()
                 if self.random_update_ua: self.session.headers.update({'User-Agent': UserAgent().random})
-            if self.auto_set_proxies:
-                try:
-                    self.session.proxies = self.proxied_session_client.getrandomproxy()
-                except Exception as err:
-                    self.logger_handle.error(f'{self.source}.get >>> {url} (Error: {err})', disable_print=self.disable_print)
-                    self.session.proxies = {}
-            else:
-                self.session.proxies = {}
+            self._autosetproxies()
             proxies = kwargs.pop('proxies', None) or self.session.proxies
-            try:
-                resp = self.session.get(url, proxies=proxies, **kwargs)
-                resp.raise_for_status()
-            except Exception as err:
-                self.logger_handle.error(f'{self.source}.get >>> {url} (Error: {err})', disable_print=self.disable_print)
-                continue
+            try: (resp := self.session.get(url, proxies=proxies, **kwargs)).raise_for_status()
+            except Exception as err: self.logger_handle.error(f'{self.source}.get >>> {url} (Error: {err}; status={getattr(locals().get("resp"), "status_code", None)})', disable_print=self.disable_print); continue
             return resp
         return resp
     '''post'''
@@ -244,21 +216,10 @@ class BaseImageClient():
             if not self.maintain_session:
                 self._initsession()
                 if self.random_update_ua: self.session.headers.update({'User-Agent': UserAgent().random})
-            if self.auto_set_proxies:
-                try:
-                    self.session.proxies = self.proxied_session_client.getrandomproxy()
-                except Exception as err:
-                    self.logger_handle.error(f'{self.source}.post >>> {url} (Error: {err})', disable_print=self.disable_print)
-                    self.session.proxies = {}
-            else:
-                self.session.proxies = {}
+            self._autosetproxies()
             proxies = kwargs.pop('proxies', None) or self.session.proxies
-            try:
-                resp = self.session.post(url, proxies=proxies, **kwargs)
-                resp.raise_for_status()
-            except Exception as err:
-                self.logger_handle.error(f'{self.source}.post >>> {url} (Error: {err})', disable_print=self.disable_print)
-                continue
+            try: (resp := self.session.post(url, proxies=proxies, **kwargs)).raise_for_status()
+            except Exception as err: self.logger_handle.error(f'{self.source}.post >>> {url} (Error: {err}; status={getattr(locals().get("resp"), "status_code", None)})', disable_print=self.disable_print); continue
             return resp
         return resp
     '''_savetopkl'''
